refactor: samples/github_repository でしか利用されていない cores/data のコードを移動
