--- packages/cores/data/lib/network.dart
@@ -1,3 +1,0 @@
-export 'src/network/json_type.dart';
-export 'src/network/provider/dio.dart';
-export 'src/network/provider/github_access_token.dart';
