+++ packages/samples/github_repository/lib/src/data/json_type.dart
