+++ packages/samples/github_repository/lib/src/data/provider/repository.dart
@@ -1,6 +1,7 @@
-import 'package:cores_data/network.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
+import 'package:samples_github_repository/src/data/json_type.dart';
+import 'package:samples_github_repository/src/data/provider/dio.dart';
 import 'package:samples_github_repository/src/domain/model/github_repository.dart';
 
 part 'repository.g.dart';
